Renaming 'vulture' package to 'heahwulf'
