+++ pom.xml
@@ -12,7 +12,7 @@
         <relativePath />
     </parent>
 
-    <groupId>fr.vulture</groupId>
+    <groupId>fr.heahwulf</groupId>
     <artifactId>hostocars</artifactId>
     <version>2.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
