+++ src/main/resources/application.properties
@@ -7,7 +7,7 @@ spring.profiles.active=dev
 logging.file.max-history=30
 logging.file.name=./log/server.log
 logging.level.com.zaxxer.hikari=OFF
-logging.level.fr.vulture.hostocars=INFO
+logging.level.fr.heahwulf.hostocars=INFO
 logging.level.io.undertow=ERROR
 logging.level.liquibase.snapshot=ERROR
 logging.level.org.hibernate=OFF
