+++ src/test/java/fr/heahwulf/hostocars/entity/CarTest.java
@@ -1,0 +1,27 @@
+package fr.heahwulf.hostocars.entity;
+
+import org.junit.jupiter.api.DisplayName;
+
+/**
+ * Test class for the {@link Car} class.
+ */
+@DisplayName("Car entity")
+class CarTest extends AbstractEntityTest<Car> {
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    Class<Car> getTestClass() {
+        return Car.class;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    Car constructTestObject() {
+        return new Car();
+    }
+
+}
