+++ src/test/java/fr/heahwulf/hostocars/entity/InterventionTest.java
@@ -1,0 +1,27 @@
+package fr.heahwulf.hostocars.entity;
+
+import org.junit.jupiter.api.DisplayName;
+
+/**
+ * Test class for the {@link Intervention} class.
+ */
+@DisplayName("Intervention entity")
+class InterventionTest extends AbstractEntityTest<Intervention> {
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    Class<Intervention> getTestClass() {
+        return Intervention.class;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    Intervention constructTestObject() {
+        return new Intervention();
+    }
+
+}
